backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
mysqlclient==2.2.1
pymysql==1.1.0
alembic==1.13.1

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Security - Rate Limiting
slowapi==0.1.9

# Email
fastapi-mail==1.4.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.8

# Logging
python-json-logger==2.0.7
